first bunch of code cleanup

+++ src/main/java/io/github/mzmine/datamodel/featuredata/FeatureDataUtils.java
@@ -144,7 +144,7 @@ public class FeatureDataUtils {
     for (int i = 1; i < series.getNumberOfValues(); i++) {
       final double thisIntensity = intensities.get(i);
       final float thisRT = scans.get(i).getRetentionTime();
-      area += (thisRT - lastRT) * (thisIntensity + lastIntensity) / 2.0 /* 60d*/;
+      area += (thisRT - lastRT) * ((float)(thisIntensity + lastIntensity)) / 2.0;
       lastIntensity = thisIntensity;
       lastRT = thisRT;
     }
